JPA repositories were implemented; Some id attribute types were changed; Deprecated Faker repositories were removed; Datasource was partly configured in application.yaml

+++ src/main/java/com/sergosoft/productservice/dto/category/CategoryCreationDto.java
@@ -39,5 +39,5 @@ public class CategoryCreationDto {
      * @implNote This field is optional.
      */
     @Positive(message = "Category parent id must be positive.")
-    Integer parentId;
+    Long parentId;
 }
